deleted unnecessary passed runParams property and added interface for RunOverviewTab
