+++ aim/web/ui/src/pages/RunDetail/RunOverviewTab/GitInfoCard/GitInfoCard.d.ts
@@ -1,0 +1,11 @@
+export interface IGitInfoCardProps {
+  data: {
+    branch: string;
+    commit: {
+      author: string;
+      hash: string;
+      timestamp: string;
+    };
+    remote_origin_url: string;
+  };
+}
