+++ aim/web/ui/src/pages/RunDetail/RunOverviewTab/RunOverViewTab.d.ts
@@ -1,11 +1,4 @@
-export interface IGitInfoCardProps {
-  data: {
-    branch: string;
-    commit: {
-      author: string;
-      hash: string;
-      timestamp: string;
-    };
-    remote_origin_url: string;
-  };
+export interface IRunOverviewTabProps {
+  runData: any;
+  runHash: string;
 }
